refactor(HelixUI): various updates
* add `HelixUI.Utils` namespace
* refactor utils to better adhere to ES6 module format
* remove redundant `src/COMPATIBILITY.md` documentation
* remove redundant initialization logic from `helix-ui.browser[.min].js`
* remove unnecessary `node-entry.js`
* normalize all import paths to use the Node ES6 format (no `.js` extensions)

BREAKING:
* rename `HelixUI.elements` namespace to `HelixUI.Elements`
* convert `HelixUI.VERSION` from an object to a string literal
