--- src/helix-ui/elements.js
@@ -1,33 +1,0 @@
-/**
- * Element Class Definitions
- * @module
- */
-export { HXAccordionElement } from './elements/HXAccordionElement.js';
-export { HXAccordionPanelElement } from './elements/HXAccordionPanelElement.js';
-export { HXAlertElement } from './elements/HXAlertElement.js';
-export { HXBusyElement } from './elements/HXBusyElement.js';
-export { HXCheckboxElement } from './elements/HXCheckboxElement.js';
-export { HXDisclosureElement } from './elements/HXDisclosureElement.js';
-export { HXDivElement } from './elements/HXDivElement.js';
-export { HXElement } from './elements/HXElement.js';
-export { HXErrorElement } from './elements/HXErrorElement.js';
-export { HXFileIconElement } from './elements/HXFileIconElement.js';
-export { HXFileInputElement } from './elements/HXFileInputElement.js';
-export { HXFileTileElement } from './elements/HXFileTileElement.js';
-export { HXIconElement } from './elements/HXIconElement.js';
-export { HXMenuElement } from './elements/HXMenuElement.js';
-export { HXMenuitemElement } from './elements/HXMenuitemElement.js';
-export { HXModalElement } from './elements/HXModalElement.js';
-export { HXPillElement } from './elements/HXPillElement.js';
-export { HXPopoverElement } from './elements/HXPopoverElement.js';
-export { HXProgressElement } from './elements/HXProgressElement.js';
-export { HXRevealElement } from './elements/HXRevealElement.js';
-export { HXSearchAssistanceElement } from './elements/HXSearchAssistanceElement.js';
-export { HXSearchElement } from './elements/HXSearchElement.js';
-export { HXTabElement } from './elements/HXTabElement.js';
-export { HXTabcontentElement } from './elements/HXTabcontentElement.js';
-export { HXTablistElement } from './elements/HXTablistElement.js';
-export { HXTabpanelElement } from './elements/HXTabpanelElement.js';
-export { HXTabsetElement } from './elements/HXTabsetElement.js';
-export { HXToastElement } from './elements/HXToastElement.js';
-export { HXTooltipElement } from './elements/HXTooltipElement.js';
