+++ src/helix-ui/elements/HXTooltipElement.js
@@ -1,5 +1,5 @@
 import { HXElement } from './HXElement';
-import { getPositionWithArrow } from '../util';
+import { getPositionWithArrow } from '../utils/position';
 import debounce from 'lodash/debounce';
 import shadowMarkup from './HXTooltipElement.html';
 import shadowStyles from './HXTooltipElement.less';
