--- src/helix-ui/util/keys.js
@@ -1,44 +1,0 @@
-/** @module */
-
-/**
- * Key/value map of key names and their keycode.
- * @enum {Integer}
- */
-const KEYS = {
-    Alt: 18,
-    /** _alias:_ `Alt` */
-    Option: 18,
-    Backspace: 8,
-    Control: 17,
-    /** _alias:_ `Control` */
-    Ctrl: 17,
-    Del: 46,
-    /** _alias:_ `Del` */
-    Delete: 46,
-    Down: 40,
-    End: 35,
-    Enter: 13,
-    /** _alias:_ `Enter` */
-    Return: 13,
-    Esc: 27,
-    /** _alias:_ `Esc` */
-    Escape: 27,
-    Home: 36,
-    Ins: 45,
-    /** _alias:_ `Ins` */
-    Insert: 45,
-    Left: 37,
-    PageDown: 34,
-    /** _alias:_ `PageDown` */
-    PgDown: 34,
-    PageUp: 33,
-    /** _alias:_ `PageUp` */
-    PgUp: 33,
-    Right: 39,
-    Shift: 16,
-    Space: 32,
-    Tab: 9,
-    Up: 38,
-};
-
-export default KEYS;
