--- src/node-entry.js
@@ -1,6 +1,0 @@
-import HelixUI from './helix-ui/index';
-import { version as VERSION } from '../package.json';
-
-HelixUI.VERSION = VERSION;
-
-export default HelixUI;
