+++ src/polyfills/index.js
@@ -1,0 +1,2 @@
+import './ChildNode';
+import './Element';
